Updated prerequisites after review
