+++ libs/application/templates/social-insurance-administration/medical-and-rehabilitation-payments/src/lib/MedicalAndRehabilitationPaymentsTemplate.ts
@@ -87,7 +87,7 @@ const MedicalAndRehabilitationPaymentsTemplate: ApplicationTemplate<
             {
               id: Roles.APPLICANT,
               formLoader: () =>
-                import('../forms/Prerequisites/index').then((module) =>
+                import('../forms/Prerequisites').then((module) =>
                   Promise.resolve(module.Prerequisites),
                 ),
               actions: [
